Doctests compiling, work on insertion function

+++ src/types.rs
@@ -1,6 +1,7 @@
 /// NodeLoc structure, representing location within octree
+#[derive(Debug)]
 pub struct NodeLoc {
-    location: Vec<u32>,
+    location: Vec<u16>,
 }
 
 impl NodeLoc {
@@ -9,7 +10,7 @@ impl NodeLoc {
     /// # Examples
     ///
     /// ```
-    /// use types::NodeLoc;
+    /// use octo::types::NodeLoc;
     ///
     /// let loc_a = NodeLoc::new((0, 0, 0));
     ///
@@ -17,24 +18,24 @@ impl NodeLoc {
     /// let loc_b = NodeLoc::new(coords);
     /// ```
     ///
-    pub fn new(coords: (u32, u32, u32)) -> NodeLoc {
+    pub fn new(coords: (u16, u16, u16)) -> NodeLoc {
         NodeLoc {
             location: vec![coords.0, coords.1, coords.2],
         }
     }
 
     /// Returns the x value of the NodeLoc
-    pub fn x(&self) -> u32 {
+    pub fn x(&self) -> u16 {
         self.location[0]
     }
 
     /// Returns the y value of the NodeLoc
-    pub fn y(&self) -> u32 {
+    pub fn y(&self) -> u16 {
         self.location[1]
     }
 
     /// Returns the z value of the NodeLoc
-    pub fn z(&self) -> u32 {
+    pub fn z(&self) -> u16 {
         self.location[2]
     }
 }
